add a stubbed out FAQ page
This is LITERALLY copied-and-pasted from the 2014 copy so it's going to
be out of date but it follows the wireframes.

One problem is that this uses jQueryUI, but the previous implementation
used some angular stuff, probably because of the need for bootstrap.

We don't have angular this time around so there may need to be some
creative fenaggling with the jQueryUI Css stuff. :P